requirements.txt
attrs==23.1.0
cattrs==23.2.3
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.46.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
openmeteo_requests==1.1.0
openmeteo_sdk==1.7.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
platformdirs==4.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
requests-cache==1.1.1
retry-requests==2.0.0
six==1.16.0
tzdata==2023.3
url-normalize==1.4.3
urllib3==2.1.0
